+++ inst/NEWS.md
@@ -1,3 +1,10 @@
+
+# rdecision 1.0.3.9000
+
+* Added option "value" to method 'set' in class 'ModVar'. This allows
+  variables to be set to an explicit value; useful in threshold finding.
+
+
 # rdecision 1.0.3
 
 * Package tests that involve sampling randomly from a distribution and
@@ -7,7 +14,7 @@
   mean and SD.
 * Added common test helpers and bespoke expectations to `testthat/setup.R`.
 * Changed vignette titles to reflect what kind of problem they illustrate,
-  rather than the problems themselves, to makes it clearer on the CRAN page.
+  rather than the problems themselves, to make it clearer on the CRAN page.
 * Added method `as_DOT` to `Graph` and `Digraph` for export to graphviz DOT
   file format to aid visualization of graphs.